+++ data-raw/correspondence_ecoicop_bh.R
@@ -5,11 +5,11 @@ library(tidyverse)
 ecoicop2bh <- readxl::read_excel("V:/SUBNATIONAL_PRICES/sources/conversions/COICOP_to_BH.xlsx",
                            sheet = "ECOICOP1_to_EUR")
 
-concordance_ecoicop2bh <- ecoicop2bh |>
+correspondence_ecoicop_bh <- ecoicop2bh |>
   select(ecoicop1_code = COICOP1,
          ecoicop1_name = NAME_EN,
          bh_code = EUR_Code,
          bh_name = EUR_name,
          type = type)
 
-usethis::use_data(concordance_ecoicop2bh, overwrite = TRUE)
+usethis::use_data(correspondence_ecoicop_bh, overwrite = TRUE)